Refactor EAGLE3
