+++ mblt_model_zoo/cli/tps.py
@@ -394,6 +394,26 @@ def _build_pipeline(
         _raise_cuda_nvml_hint(e)
 
 
+def _extract_eagle3_pipeline_kwargs(args: argparse.Namespace) -> dict[str, Any]:
+    """Return EAGLE-3 specific pipeline kwargs from parsed CLI arguments."""
+    return {
+        "base_embedding_path": args.base_embedding_path,
+        "draft_embedding_path": args.draft_embedding_path,
+        "base_mxq_path": args.base_mxq_path,
+        "draft_mxq_path": args.draft_mxq_path,
+        "fc_mxq_path": args.fc_mxq_path,
+        "base_core_mode": args.base_core_mode,
+        "draft_core_mode": args.draft_core_mode,
+        "fc_core_mode": args.fc_core_mode,
+        "base_target_cores": args.base_target_cores,
+        "draft_target_cores": args.draft_target_cores,
+        "fc_target_cores": args.fc_target_cores,
+        "base_target_clusters": args.base_target_clusters,
+        "draft_target_clusters": args.draft_target_clusters,
+        "fc_target_clusters": args.fc_target_clusters,
+    }
+
+
 def _iter_rows_for_csv(result: Any) -> Iterable[dict[str, Any]]:
     def _optional_values(values: Sequence[Any], length: int) -> list[Any]:
         return [values[idx] if idx < len(values) else None for idx in range(length)]
@@ -785,22 +805,9 @@ def _run_text_measure(args: argparse.Namespace) -> int:
         device_map=args.device_map,
         revision=args.revision,
         embedding_weight=args.embedding_weight,
-        base_embedding_path=args.base_embedding_path,
-        draft_embedding_path=args.draft_embedding_path,
         mxq_path=args.mxq_path,
-        base_mxq_path=args.base_mxq_path,
-        draft_mxq_path=args.draft_mxq_path,
-        fc_mxq_path=args.fc_mxq_path,
         core_mode=args.core_mode,
-        base_core_mode=args.base_core_mode,
-        draft_core_mode=args.draft_core_mode,
-        fc_core_mode=args.fc_core_mode,
-        base_target_cores=args.base_target_cores,
-        draft_target_cores=args.draft_target_cores,
-        fc_target_cores=args.fc_target_cores,
-        base_target_clusters=args.base_target_clusters,
-        draft_target_clusters=args.draft_target_clusters,
-        fc_target_clusters=args.fc_target_clusters,
+        **_extract_eagle3_pipeline_kwargs(args),
         target_cores=args.target_cores,
         target_clusters=args.target_clusters,
     )
@@ -1050,22 +1057,9 @@ def _run_vlm_measure(args: argparse.Namespace) -> int:
         device_map=args.device_map,
         revision=args.revision,
         embedding_weight=args.embedding_weight,
-        base_embedding_path=args.base_embedding_path,
-        draft_embedding_path=args.draft_embedding_path,
         mxq_path=args.mxq_path,
-        base_mxq_path=args.base_mxq_path,
-        draft_mxq_path=args.draft_mxq_path,
-        fc_mxq_path=args.fc_mxq_path,
         core_mode=args.core_mode,
-        base_core_mode=args.base_core_mode,
-        draft_core_mode=args.draft_core_mode,
-        fc_core_mode=args.fc_core_mode,
-        base_target_cores=args.base_target_cores,
-        draft_target_cores=args.draft_target_cores,
-        fc_target_cores=args.fc_target_cores,
-        base_target_clusters=args.base_target_clusters,
-        draft_target_clusters=args.draft_target_clusters,
-        fc_target_clusters=args.fc_target_clusters,
+        **_extract_eagle3_pipeline_kwargs(args),
         target_cores=args.target_cores,
         target_clusters=args.target_clusters,
     )
@@ -1291,22 +1285,9 @@ def _run_text_sweep(args: argparse.Namespace) -> int:
         device_map=args.device_map,
         revision=args.revision,
         embedding_weight=args.embedding_weight,
-        base_embedding_path=args.base_embedding_path,
-        draft_embedding_path=args.draft_embedding_path,
         mxq_path=args.mxq_path,
-        base_mxq_path=args.base_mxq_path,
-        draft_mxq_path=args.draft_mxq_path,
-        fc_mxq_path=args.fc_mxq_path,
         core_mode=args.core_mode,
-        base_core_mode=args.base_core_mode,
-        draft_core_mode=args.draft_core_mode,
-        fc_core_mode=args.fc_core_mode,
-        base_target_cores=args.base_target_cores,
-        draft_target_cores=args.draft_target_cores,
-        fc_target_cores=args.fc_target_cores,
-        base_target_clusters=args.base_target_clusters,
-        draft_target_clusters=args.draft_target_clusters,
-        fc_target_clusters=args.fc_target_clusters,
+        **_extract_eagle3_pipeline_kwargs(args),
         target_cores=args.target_cores,
         target_clusters=args.target_clusters,
     )
@@ -1777,22 +1758,9 @@ def _run_vlm_sweep(args: argparse.Namespace) -> int:
         device_map=args.device_map,
         revision=args.revision,
         embedding_weight=args.embedding_weight,
-        base_embedding_path=args.base_embedding_path,
-        draft_embedding_path=args.draft_embedding_path,
         mxq_path=args.mxq_path,
-        base_mxq_path=args.base_mxq_path,
-        draft_mxq_path=args.draft_mxq_path,
-        fc_mxq_path=args.fc_mxq_path,
         core_mode=args.core_mode,
-        base_core_mode=args.base_core_mode,
-        draft_core_mode=args.draft_core_mode,
-        fc_core_mode=args.fc_core_mode,
-        base_target_cores=args.base_target_cores,
-        draft_target_cores=args.draft_target_cores,
-        fc_target_cores=args.fc_target_cores,
-        base_target_clusters=args.base_target_clusters,
-        draft_target_clusters=args.draft_target_clusters,
-        fc_target_clusters=args.fc_target_clusters,
+        **_extract_eagle3_pipeline_kwargs(args),
         target_cores=args.target_cores,
         target_clusters=args.target_clusters,
     )
